+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Supported `beginCreate(Context)` in `VirtualMachine`.
+
 ### Breaking Changes
 
 ### Bugs Fixed